requirements.txt
# Data Collection & Processing
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
requests>=2.31.0
fredapi>=0.5.0  

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Statistical Analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Utilities
PyYAML>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
openpyxl>=3.1.0
python-docx>=1.0.0